requirements.txt
selenium
python-dotenv
discordwebhook
